# compiler: allow inlining of internal/asm package functions