Go uses a scoring mechanism to determine which functions get inlined. Each statement gets assigned a scope and anything with a total below or equal to 80 gets inlined.

This PR modifies the `asm.CodeSegment` and `asm.Buffer` types to reduce the cost of functions that were on the hot path and allow inlining.

Chasing down inlining scores is somewhat of a dark art, but I was able to get the function scores by embedding `*CodeSegment` in `Buffer` so the call sites can see the inline-able leaf functions of `CodeSegment`, and the API remains identical from the standpoint of `Buffer`.

For the record, here's one way to ask the Go compiler to show functions that cannot be inlined in a package:
```
$ go build -gcflags '-m -m' ./internal/asm/ 2>&1 | grep 'cannot inline'
internal/asm/buffer.go:48:6: cannot inline (*CodeSegment).Map: function too complex: cost 153 exceeds budget 80
internal/asm/buffer.go:66:6: cannot inline (*CodeSegment).Unmap: function too complex: cost 111 exceeds budget 80
internal/asm/buffer.go:184:6: cannot inline (*CodeSegment).Grow: function too complex: cost 114 exceeds budget 80
internal/asm/buffer.go:175:6: cannot inline (*CodeSegment).growToSize: marked go:noinline
internal/asm/buffer.go:148:6: cannot inline (*CodeSegment).AppendBytes: function too complex: cost 86 exceeds budget 80
internal/asm/buffer.go:157:6: cannot inline (*CodeSegment).AppendUint32: function too complex: cost 82 exceeds budget 80
internal/asm/impl.go:42:6: cannot inline (*BaseAssemblerImpl).FinalizeJumpTableEntry: function too complex: cost 313 exceeds budget 80
```


This shrinks a couple of percentage points on CPU benchmarks, for example:
```
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
                                 │ /tmp/bench.1 │            /tmp/bench.2            │
                                 │    sec/op    │   sec/op     vs base               │
Compilation/without_extern_cache    5.051m ± 2%   4.899m ± 1%  -3.02% (p=0.005 n=12)

                                 │ /tmp/bench.1 │          /tmp/bench.2          │
                                 │     B/op     │     B/op      vs base          │
Compilation/without_extern_cache   597.8Ki ± 0%   597.8Ki ± 0%  ~ (p=0.291 n=12)

                                 │ /tmp/bench.1 │         /tmp/bench.2          │
                                 │  allocs/op   │  allocs/op   vs base          │
Compilation/without_extern_cache    1.074k ± 0%   1.074k ± 0%  ~ (p=1.000 n=12)
```
on darwin/amd64 the change did not seem to have any effect, so I avoided a few unsafe optimizations in favor of safer code (see comment in `AppendUint32`):
```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                 │ /tmp/bench.1 │         /tmp/bench.2          │
                                 │    sec/op    │   sec/op     vs base          │
Compilation/without_extern_cache    1.850m ± 2%   1.819m ± 2%  ~ (p=0.178 n=12)

                                 │ /tmp/bench.1 │          /tmp/bench.2          │
                                 │     B/op     │     B/op      vs base          │
Compilation/without_extern_cache   565.6Ki ± 0%   565.6Ki ± 0%  ~ (p=0.640 n=12)

                                 │ /tmp/bench.1 │          /tmp/bench.2          │
                                 │  allocs/op   │ allocs/op   vs base            │
Compilation/without_extern_cache     934.0 ± 0%   934.0 ± 0%  ~ (p=1.000 n=12) ¹
```